[157] small fix & mapper added (draft)

+++ dao/src/main/java/greencity/entity/Address.java
@@ -4,6 +4,8 @@ import jakarta.persistence.*;
 import jakarta.validation.constraints.NotNull;
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Entity
 @NoArgsConstructor
 @AllArgsConstructor
@@ -22,31 +24,20 @@ public class Address {
     private EventDateLocation eventDateLocation;
 
     @NotNull
-    private Double latitude;
+    private BigDecimal latitude;
 
     @NotNull
-    private Double longitude;
+    private BigDecimal longitude;
 
     private String countryEn;
-
     private String countryUa;
-
     private String regionEn;
-
     private String regionUa;
-
     private String cityEn;
-
     private String cityUa;
-
     private String streetEn;
-
     private String streetUa;
-
     private String houseNumber;
-
     private String formattedAddressEn;
-
     private String formattedAddressUa;
-
 }
